imagetext/requirements/testing.txt
-r common.txt
autopep8==1.4.4
coverage==5.1
django-debug-toolbar==2.0
flake8==3.7.7
flake8-blind-except==0.1.1
flake8-docstrings==1.5.0
freezegun==0.3.15
isort==4.3.4
mccabe==0.6.1
mock==4.0.2
pep257==0.7.0
pep8==1.7.1
pep8-naming==0.4.1
pycodestyle==2.5.0
pydocstyle==5.0.2